Feature to panic on weight overflow

+++ rust/template/differential_datalog/src/api/update_handler.rs
@@ -134,9 +134,10 @@ impl<F: Callback> UpdateHandler for CallbackUpdateHandler<F> {
 }
 
 impl<F: Callback> MTUpdateHandler for CallbackUpdateHandler<F> {
+    #[allow(clippy::useless_conversion)]
     fn mt_update_cb(&self) -> Arc<dyn RelationCallback> {
         let cb = self.cb.clone();
-        Arc::new(move |relid, v, w| cb(relid, &v.clone().into_record(), w as isize))
+        Arc::new(move |relid, v, w| cb(relid, &v.clone().into_record(), i64::from(w) as isize))
     }
 }
 
@@ -182,13 +183,19 @@ impl UpdateHandler for ExternCUpdateHandler {
 
 #[cfg(feature = "c_api")]
 impl MTUpdateHandler for ExternCUpdateHandler {
+    #[allow(clippy::useless_conversion)]
     fn mt_update_cb(&self) -> Arc<dyn RelationCallback> {
         let cb = self.cb;
         let cb_arg = self.cb_arg;
 
         Arc::new(move |relid, v, w| {
             let value = v.clone().into_record();
-            cb(cb_arg, relid, &value as *const Record, w as isize);
+            cb(
+                cb_arg,
+                relid,
+                &value as *const Record,
+                i64::from(w) as isize,
+            );
         })
     }
 }
@@ -216,9 +223,10 @@ impl UpdateHandler for MTValMapUpdateHandler {
 }
 
 impl MTUpdateHandler for MTValMapUpdateHandler {
+    #[allow(clippy::useless_conversion)]
     fn mt_update_cb(&self) -> Arc<dyn RelationCallback> {
         let db = self.db.clone();
-        Arc::new(move |relid, v, w| db.lock().unwrap().update(relid, v, w as isize))
+        Arc::new(move |relid, v, w| db.lock().unwrap().update(relid, v, i64::from(w) as isize))
     }
 }
 
@@ -540,6 +548,7 @@ impl UpdateHandler for ThreadUpdateHandler {
 }
 
 impl MTUpdateHandler for ThreadUpdateHandler {
+    #[allow(clippy::useless_conversion)]
     fn mt_update_cb(&self) -> Arc<dyn RelationCallback> {
         let channel = self.msg_channel.clone();
         Arc::new(move |relid, v, w| {
@@ -547,7 +556,7 @@ impl MTUpdateHandler for ThreadUpdateHandler {
                 .send(Msg::Update {
                     relid,
                     v: v.clone(),
-                    w: w as isize,
+                    w: i64::from(w) as isize,
                 })
                 .unwrap();
         })
